+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Change Log
 
-## dev-master
+## dev-master (2017-10-08)
+- Added support for HTTP authentication in the request.
+- Change unitTest.
+- Refactoring.
+
+## dev-master (2017-10-01)
+- Added support for working with HTTP headers.
+- Added unitTest for class Brownie\HttpClient\Headers.
+- Refactoring.
+
+## dev-master (2017-09-27)
 Features:
 - Initinal commit.
 - Added HTTP Client method "request".
@@ -11,5 +21,3 @@ Features:
 - Added unitTest for class Brownie\HttpClient\Exception\ValidateException.
 - Added unitTest for class Brownie\HttpClient\Request.
 - Added unitTest for class Brownie\HttpClient\Response.
-- Added unitTest for class Brownie\HttpClient\Headers.
-- Refactoring.